Callback into FanControl after differential pressure read
This callback will be used to run differential pressure calibration.

Further, timing for fan measurement and for DP measurement was changed
to 1s + 10ms to ensure that PID regulator will surely pick the new value.
Previously, there was a chance that PID regulator wouldn't run for one
loop due to exact timing of 1s.